requirements.txt
torch
torchvision
matplotlib
numpy
pandas
scikit-learn
Pillow
tqdm
opencv-python
ipykernel
